Feat: 메인페이지와 랜딩페이지 분리하기
session의 유무를 판단해 페이지를 분리하는 것이 아닌 middleware를 통해
토큰의 유무 확인 후 리다이렉트되는 방식으로 변경
