+++ src/app/(landing)/page.tsx
@@ -1,0 +1,7 @@
+import LandingSection from "@/templates/LandingSection";
+
+const LandingPage = () => {
+  return <LandingSection />;
+};
+
+export default LandingPage;
