+++ src/app/(main)/my/page.tsx
@@ -1,0 +1,7 @@
+import MainSection from "@/templates/MainSection";
+
+const HomePage = async () => {
+  return <MainSection />;
+};
+
+export default HomePage;
